--- a/benchmark/script.js
+++ b/benchmark/script.js
@@ -5,7 +5,7 @@ var bitcore = require('..');
 var async = require('async');
 var blockData = require('./block-357238.json');
 
-var maxTime = 10;
+var maxTime = 30;
 
 console.log('Benchmarking Script');
 console.log('---------------------------------------');