--- a/lib/encoding/bufferreader.js
+++ b/lib/encoding/bufferreader.js
@@ -22,7 +22,7 @@ var BufferReader = function BufferReader(buf) {
       throw new TypeError('Invalid hex string');
 
     this.set({
-      buf: b,
+      buf: Buffer.from(buf, 'hex'),
     });
   } else if (_.isObject(buf)) {
     var obj = buf;
@@ -182,7 +182,7 @@ BufferReader.prototype.readVarintBN = function() {
 };
 
 BufferReader.prototype.reverse = function() {
-  var buf = new Buffer(this.buf.length);
+  var buf = Buffer.alloc(this.buf.length);
   for (var i = 0; i < buf.length; i++) {
     buf[i] = this.buf[this.buf.length - 1 - i];
   }