--- a/lib/encoding/bufferreader.js
+++ b/lib/encoding/bufferreader.js
@@ -17,12 +17,12 @@ var BufferReader = function BufferReader(buf) {
       buf: buf
     });
   } else if (_.isString(buf)) {
-    var b = new Buffer(buf, 'hex');
+    var b = Buffer.from(buf);
     if (b.length * 2 != buf.length) 
       throw new TypeError('Invalid hex string');
 
     this.set({
-      buf: Buffer.from(buf, 'hex'),
+      buf: b
     });
   } else if (_.isObject(buf)) {
     var obj = buf;