--- a/lib/encoding/bufferreader.js
+++ b/lib/encoding/bufferreader.js
@@ -17,7 +17,7 @@ var BufferReader = function BufferReader(buf) {
       buf: buf
     });
   } else if (_.isString(buf)) {
-    var b = Buffer.from(buf);
+    var b = Buffer.from(buf, 'hex');
     if (b.length * 2 != buf.length) 
       throw new TypeError('Invalid hex string');