--- a/lib/transaction/sighash.js
+++ b/lib/transaction/sighash.js
@@ -202,7 +202,7 @@ var sighash = function sighash(transaction, sighashType, inputNumber, subscript,
     // The SIGHASH_SINGLE bug.
     // https://bitcointalk.org/index.php?topic=260595.0
     if (inputNumber >= txcopy.outputs.length) {
-      return new Buffer(SIGHASH_SINGLE_BUG, 'hex');
+      return Buffer.from(SIGHASH_SINGLE_BUG, 'hex');
     }
 
     txcopy.outputs.length = inputNumber + 1;