--- a/lib/transaction/transaction.js
+++ b/lib/transaction/transaction.js
@@ -102,7 +102,8 @@ var hashProperty = {
   configurable: false,
   enumerable: true,
   get: function() {
-    return new BufferReader(this._getHash()).readReverse().toString('hex');
+    this._hash = new BufferReader(this._getHash()).readReverse().toString('hex');
+    return this._hash;
   }
 };
 Object.defineProperty(Transaction.prototype, 'hash', hashProperty);