* new Buffer() is deprecated
  * https://nodejs.org/api/buffer.html#buffer_new_buffer_array
* new Buffer([]) should be Buffer.from([])
* new Buffer(9) should be Buffer.alloc(9)
* BufferWriter should keep track of length as you write
* Transaction.hash getter should cache the last hash, until you change it.


Before Changes 
```
Benchmarking Script
---------------------------------------
isPublicKeyHashIn x 42,112,577 ops/sec ±2.14% (421 runs sampled)
isPublicKeyOut x 30,459,210 ops/sec ±2.80% (430 runs sampled)
toAddress x 65,897 ops/sec ±3.15% (323 runs sampled)
getAddressInfo x 86,997 ops/sec ±2.68% (325 runs sampled)
Done
----------------------------------------------------------------------
Finished
```


After Changes 
```
Benchmarking Script
---------------------------------------
isPublicKeyHashIn x 65,906,640 ops/sec ±0.98% (516 runs sampled)
isPublicKeyOut x 51,775,114 ops/sec ±0.92% (510 runs sampled)
toAddress x 122,002 ops/sec ±1.03% (437 runs sampled)
getAddressInfo x 143,790 ops/sec ±0.93% (528 runs sampled)
Done
----------------------------------------------------------------------
Finished
```