@matiu 2018-02-13
There is a failing test at:
https://travis-ci.org/bitpay/bitcore-lib/builds/340632730#L1213

Could you check it?